MYPYCHARM_CODES/requirements.txt
fastapi
firebase-admin
streamlit
Pillow
